+++ ReleaseNotes.md
@@ -2,5 +2,5 @@ Here's a list of the latest changes:
 
 **3.33.0**
 
-- [Change - Adding original ticket price to guests when downloading the guest list report]()
-
+- [Fix - Adding original ticket price to guests when downloading the guest list report](https://trello.com/c/7boFqvit/663-show-the-ticket-price-in-the-download-guests)
+- [Fix - Ensure that we display the same date format in all the places](https://trello.com/c/53uLx8Rs/645-fix-all-the-display-dates-to-be-consistent)